pyproject.toml
[project]
name = "recast-bench"
version = "0.1.0"
description = "ReCast: A benchmark for evaluating LLM causal reasoning on real-world academic text"
readme = "README.md"
license = { text = "CC-BY-4.0" }
requires-python = ">=3.10"
authors = [
    { name = "Ryan Saklad" },
    { name = "Aman Chadha" },
    { name = "Oleg Pavlov" },
    { name = "Raha Moraffah" },
]

dependencies = [
    "datasets>=2.14.0",
    "networkx>=3.0",
    "numpy>=1.24.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
    "tabulate>=0.9.0",
    "tqdm>=4.65.0",
    "pandas>=2.0.0",
    "huggingface_hub>=0.19.0",
    "pyyaml>=6.0.0",
    "cdt>=0.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]
teagim = [
    "torch-geometric>=2.4.0",
    "tqdm>=4.65.0",
    "pandas>=2.0.0",
    "seaborn>=0.12.0",
    "matplotlib>=3.7.0",
]

[project.scripts]
recast = "recast.cli:main"

[project.urls]
Homepage = "https://github.com/RyanSaklad/ReCast"
Repository = "https://github.com/RyanSaklad/ReCast"
Paper = "https://arxiv.org/abs/2505.18931"
Dataset = "https://huggingface.co/datasets/RyanSaklad/ReCast"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
